Move const.h and add decls.h
Moved to 'common' directory. Added 'decls.h' to declare common types more specifically to avoid code verbosity.
