+++ src/RePak.vcxproj
@@ -32,13 +32,14 @@
   <ItemGroup>
     <ClInclude Include="application\repak.h" />
     <ClInclude Include="assets\assets.h" />
+    <ClInclude Include="common\const.h" />
+    <ClInclude Include="common\decls.h" />
     <ClInclude Include="logic\pakfile.h" />
     <ClInclude Include="logic\rmem.h" />
     <ClInclude Include="logic\rtech.h" />
     <ClInclude Include="math\color.h" />
     <ClInclude Include="math\vector.h" />
     <ClInclude Include="pch.h" />
-    <ClInclude Include="public\const.h" />
     <ClInclude Include="public\material.h" />
     <ClInclude Include="public\rpak.h" />
     <ClInclude Include="public\studio.h" />
