+++ src/common/decls.h
@@ -1,0 +1,19 @@
+#pragma once
+
+// Common type declarations to reduce code verbosity.
+using std::pair;
+using std::regex;
+using std::string;
+using std::wstring;
+using std::u16string;
+using std::u32string;
+using std::vector;
+using std::fstream;
+using std::ifstream;
+using std::ofstream;
+using std::stringstream;
+using std::ostringstream;
+using std::unordered_map;
+
+namespace fs = std::filesystem;
+namespace js = rapidjson;
